structured Rust project a little better
