+++ src-tauri/src/commands/collection.rs
@@ -1,12 +1,11 @@
+use crate::Pids;
 use std::io;
-use std::path::{Path, PathBuf, MAIN_SEPARATOR};
+use std::path::{Path, PathBuf};
 use std::sync::{Arc, Mutex};
 use std::u64;
 use std::vec;
 use tauri::State;
 use tokio::fs;
-
-use crate::Pids;
 
 #[derive(serde::Serialize)]
 #[allow(non_snake_case)]
@@ -334,38 +333,4 @@ fn add_asset(asset: &PathBuf, work: &mut Work) {
             });
         };
     };
-}
-
-#[tauri::command]
-pub async fn read_collection_list(
-    collection_path: String,
-    list_name: String,
-) -> Option<Vec<String>> {
-    let path = format!("{collection_path}{MAIN_SEPARATOR}.{list_name}");
-
-    if let Ok(list_contents) = fs::read_to_string(path).await {
-        let lines = list_contents.lines().map(|line| line.to_string()).collect();
-
-        Some(lines)
-    } else {
-        None
-    }
-}
-
-#[tauri::command]
-pub async fn write_collection_list(
-    collection_path: String,
-    list_name: String,
-    list: Vec<String>,
-) -> Result<(), String> {
-    let path = format!("{collection_path}{MAIN_SEPARATOR}.{list_name}");
-
-    fs::write(path, list.join("\n"))
-        .await
-        .map_err(|error| error.to_string())
-}
-
-#[tauri::command]
-pub fn open_external(path_or_url: String) -> Result<(), String> {
-    opener::open(path_or_url).map_err(|error| error.to_string())
 }
