+++ src-tauri/src/commands/mod.rs
@@ -1,0 +1,7 @@
+mod collection;
+mod collection_list;
+mod utils;
+
+pub use collection::*;
+pub use collection_list::*;
+pub use utils::*;
