+++ src-tauri/src/commands/utils.rs
@@ -1,0 +1,4 @@
+#[tauri::command]
+pub fn open_external(path_or_url: String) -> Result<(), String> {
+    opener::open(path_or_url).map_err(|error| error.to_string())
+}
