+++ src-tauri/src/main.rs
@@ -1,9 +1,9 @@
 // Prevents additional console window on Windows in release, DO NOT REMOVE!!
 #![cfg_attr(not(debug_assertions), windows_subsystem = "windows")]
 
-use std::sync::{Arc, Mutex};
+mod commands;
 
-mod handlers;
+use std::sync::{Arc, Mutex};
 
 pub struct Pids {
     read_collection_pid: Arc<Mutex<usize>>,
@@ -15,10 +15,10 @@ fn main() {
             read_collection_pid: Arc::new(Mutex::new(0)),
         })
         .invoke_handler(tauri::generate_handler![
-            handlers::read_collection,
-            handlers::read_collection_list,
-            handlers::write_collection_list,
-            handlers::open_external
+            commands::read_collection,
+            commands::read_collection_list,
+            commands::write_collection_list,
+            commands::open_external
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
